LPD-48736 Merge UnsatisfiedComponentOSGiCommands.java and SoftCircularDependencyOSGiCommands.java into one class
